Add router control, admin view and login view
