+++ package.json
@@ -13,6 +13,7 @@
     "firebase": "^11.7.1",
     "moment": "^2.30.1",
     "vue": "^3.5.13",
+    "vue-router": "^4.5.1",
     "vuetify": "^3.8.4"
   },
   "devDependencies": {
