+++ src/main.js
@@ -1,6 +1,7 @@
 import { createApp } from 'vue'
 import './style.css'
 import App from './App.vue'
+import router from './router';
 
 import VueDatePicker from '@vuepic/vue-datepicker';
 import '@vuepic/vue-datepicker/dist/main.css'
@@ -11,6 +12,8 @@ import vuetify from './plugins/vuetify';
 const app = createApp(App);
 
 app.component('VueDatePicker', VueDatePicker);
+
+app.use(router);
 
 app.use(vuetify);
