+++ src/style.css
@@ -49,21 +49,10 @@ button:focus-visible {
 }
 
 #app {
+  width: 100%;;
   max-width: 1280px;
   margin: 0 auto;
   text-align: center;
-}
-
-@media (max-width: 960px) {
-  #app {
-    padding: 0 2em 2em 2em;
-  }
-}
-
-@media (min-width: 961px) {
-  #app {
-    padding: 0 2em;
-  }
 }
 
 @media (prefers-color-scheme: light) {
